Marabunta versioning improvements (#32)
* First version is now called "setup"

Rationale:
"""
Currently Marabunta migration file starts with a version that we usually set as 10.0.0 / 11.0.0, ...

This is pretty confusing for 2 reasons: when we start we don't really have a release yet, but above all, this version is the one that install/setups the project when the containers starts on a new DB. The further ones are upgrades.
This is also the place where we should squash every additions done in later upgrades.

If a migration.yml file does not contain any version named "setup", print a warning. If the setup is not the first of the list of versions, raise an error.
"""

* Support 5 digits version

Rationale:
""""
The versions Marabunta is using are the ones in 'distutils.version.StrictVersion'.

Examples: 0.4.0 - 0.4.1 - 0.5a1 - 0.5b3 - 0.5 - 0.9.6 - 1.0 - 1.0.4a3 - 1.0.4b1 - 1.0.4

The versions we use for our projects are:  <odooversion>.<feature>.<patch> (10.0.0, 11.1.2). Which leaves us with only 2 digits as the first one is used for the Odoo version.

We have no version available to indicate when our project is in production, or maybe that we are in a large "2nd phase". We'd like to have <major>.<feature>.<patch>, organized as:

major: 1 when in production, can be incremented to represent large evolutions / rework
feature: same than currently, new features, ...
patch: correction of a deployed major/minor version (hotfix)
However we want to keep the odoo version, so 11.0.0 becomes 11.0.0.0.0, 11.1.0 becomes 11.0.0.1.0... (why not 11.0.1.0? to be consistent with addons version scheme)
"""