Run backup commands in a shell
The initial implementation of the backups do not allow to inject
environment variables in the commands. For instance, with the following
block:

    backup:
      command: echo
      args: "backup command on ${DB_NAME}"
      stop_on_failure: true
      ignore_if: test "${RUNNING_ENV}" != "prod"

ignore_if would really compare the "${RUNNING_ENV}" chain without
replacing it by the environment variable's value. Similarly, the DB_NAME
in the command args would be printed the exact same way without
interpolation.

Now, the operations can be wrapped in a shell (sh), which will take care
of the interpolation. The version operations are still run without
a shell (although you can use 'sh -c' in the yaml file if you need), but
now the backup commands use one.

This change has few effects on the existing behavior:

* pexpect.spawn can and is documented to use a string for the command +
args, so we no longer split them as a list, this is also safer as
splitting can be tricky in some cases
* as a result of no longer splitting the command, the output may change
a bit, for instance "echo 'foo bar'" was printed as "echo foo bar" in
the output, now it's "echo 'foo bar'"
* The backup options in the yaml file no longer expect command and args
in separate keys, only "command" is expected:

    backup:
      command: echo "backup command on ${DB_NAME}"
      stop_on_failure: true
      ignore_if: test "${RUNNING_ENV}" != "prod"