Add option for 503 urls while migrating
During a migration a web server is started that returns an information
page to all requested urls on the domain.

For an Odoo instance with some controllers this is not optimal, because
if a web service is called during the migration the information page
which is a 200 HTTTP code could be erroneously interpreted by the calling
party.

To resolve this a new option is available, it accepts a comma separated
list of urls path for whose a 503 HTTP code will be returned instead of
the default information page.