[IMP] new option 'web-healthcheck-path'
This option can be used to set a monitoring healthcheck path
that the web server of Marabunta will take into account
to return a HTTP 200 instead of the usual 503.

The platform managing the Odoo/Marabunta service could then flag
it as "healthy" and continue to formard HTTP requests to it.